argh, 'fixed' the wrong argument
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1548950 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/utils/CRC32VerifyingInputStream.java
@@ -28,7 +28,7 @@ import java.util.zip.CRC32;
  */
 public class CRC32VerifyingInputStream extends ChecksumVerifyingInputStream {
     
-    public CRC32VerifyingInputStream(final InputStream in, final int size, final long expectedCrc32) {
+    public CRC32VerifyingInputStream(final InputStream in, final long size, final int expectedCrc32) {
         super(new CRC32(), in, size, expectedCrc32);
     }
